feat: Refactor data source display in cards. Dynamically load data source in navbar.

+++ backend/src/main/java/fr/inra/urgi/faidare/elasticsearch/repository/ESSuggestRepository.java
@@ -14,6 +14,6 @@ public interface ESSuggestRepository<C extends PaginationCriteria> {
     /**
      * Suggest field values
      */
-    LinkedHashSet<String> suggest(String field, String searchText, Long fetchSize, C criteria);
+    LinkedHashSet<String> suggest(String field, String searchText, Integer fetchSize, C criteria);
 
 }
